# Monthly archiving should use the smaller amount of weeks and days inside the month