--- a/plugins/UserCountry/tests/System/AttributeHistoricalDataWithLocationsTest.php
+++ b/plugins/UserCountry/tests/System/AttributeHistoricalDataWithLocationsTest.php
@@ -10,9 +10,11 @@ namespace Piwik\Plugins\UserCountry\Test\Integration;
 
 use Piwik\Common;
 use Piwik\Db;
+use Piwik\Plugin;
 use Piwik\Plugins\UserCountry\Commands\AttributeHistoricalDataWithLocations;
 use Piwik\Tests\Fixtures\ManyVisitsWithGeoIP;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
+use Piwik\Translate;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
 
@@ -101,6 +103,12 @@ class AttributeHistoricalDataWithLocationsTest extends IntegrationTestCase
             'period'  => 'month'
         );
 
+        // we need to manually reload the translations since they get reset for some reason in IntegrationTestCase::tearDown();
+        // if we do not load translations, a DataTable\Map containing multiple periods will contain only one DataTable having
+        // the label `General_DateRangeFromTo` instead of many like `From 2010-01-04 to 2010-01-11`, ' `From 2010-01-11 to 2010-01-18`
+        // As those data tables would all have the same prettyfied period label they would overwrite each other.
+        Translate::loadAllTranslations();
+
         $this->assertApiResponseEqualsExpected("UserCountry.getCountry", $queryParams);
         $this->assertApiResponseEqualsExpected("UserCountry.getContinent", $queryParams);
         $this->assertApiResponseEqualsExpected("UserCountry.getRegion", $queryParams);
@@ -136,6 +144,7 @@ class AttributeHistoricalDataWithLocationsTest extends IntegrationTestCase
 
     public static function configureFixture($fixture)
     {
+        $fixture->loadTranslations    = true;
         // empty (undo IntegrationTestCase configuring)
     }