--- a/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getCity_month.xml
+++ b/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getCity_month.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <result>
 	<row>
-		<label>General_Unknown</label>
+		<label>Unknown</label>
 		<nb_visits>22</nb_visits>
 		<nb_actions>39</nb_actions>
 		<max_actions>3</max_actions>
@@ -23,16 +23,16 @@
 		<revenue>165</revenue>
 		<sum_daily_nb_uniq_visitors>11</sum_daily_nb_uniq_visitors>
 		<sum_daily_nb_users>0</sum_daily_nb_users>
-		<city_name>General_Unknown</city_name>
+		<city_name>Unknown</city_name>
 		<city>xx</city>
 		<region>xx</region>
 		<country>xx</country>
-		<country_name>General_Unknown</country_name>
-		<region_name>General_Unknown</region_name>
+		<country_name>Unknown</country_name>
+		<region_name>Unknown</region_name>
 		<logo>plugins/UserCountry/images/flags/xx.png</logo>
 	</row>
 	<row>
-		<label>Vancouver, British Columbia, UserCountry_country_ca</label>
+		<label>Vancouver, British Columbia, Canada</label>
 		<nb_visits>6</nb_visits>
 		<nb_actions>10</nb_actions>
 		<max_actions>3</max_actions>
@@ -60,12 +60,12 @@
 		<city_name>Vancouver</city_name>
 		<region>BC</region>
 		<country>ca</country>
-		<country_name>UserCountry_country_ca</country_name>
+		<country_name>Canada</country_name>
 		<region_name>British Columbia</region_name>
 		<logo>plugins/UserCountry/images/flags/ca.png</logo>
 	</row>
 	<row>
-		<label>Besançon, Franche-Comte, UserCountry_country_fr</label>
+		<label>Besançon, Franche-Comte, France</label>
 		<nb_visits>3</nb_visits>
 		<nb_actions>5</nb_actions>
 		<max_actions>3</max_actions>
@@ -93,12 +93,12 @@
 		<city_name>Besançon</city_name>
 		<region>A6</region>
 		<country>fr</country>
-		<country_name>UserCountry_country_fr</country_name>
+		<country_name>France</country_name>
 		<region_name>Franche-Comte</region_name>
 		<logo>plugins/UserCountry/images/flags/fr.png</logo>
 	</row>
 	<row>
-		<label>Lhasa, General_Unknown, UserCountry_country_ti</label>
+		<label>Lhasa, Unknown, Tibet</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>3</nb_actions>
 		<max_actions>2</max_actions>
@@ -126,12 +126,12 @@
 		<city_name>Lhasa</city_name>
 		<region>1</region>
 		<country>ti</country>
-		<country_name>UserCountry_country_ti</country_name>
-		<region_name>General_Unknown</region_name>
+		<country_name>Tibet</country_name>
+		<region_name>Unknown</region_name>
 		<logo>plugins/UserCountry/images/flags/ti.png</logo>
 	</row>
 	<row>
-		<label>Rome, Lazio, UserCountry_country_it</label>
+		<label>Rome, Lazio, Italy</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>4</nb_actions>
 		<max_actions>3</max_actions>
@@ -159,7 +159,7 @@
 		<city_name>Rome</city_name>
 		<region>07</region>
 		<country>it</country>
-		<country_name>UserCountry_country_it</country_name>
+		<country_name>Italy</country_name>
 		<region_name>Lazio</region_name>
 		<logo>plugins/UserCountry/images/flags/it.png</logo>
 	</row>