--- a/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getCountry_month.xml
+++ b/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getCountry_month.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <result>
 	<row>
-		<label>UserCountry_country_cn</label>
+		<label>China</label>
 		<nb_visits>18</nb_visits>
 		<nb_actions>32</nb_actions>
 		<max_actions>3</max_actions>
@@ -30,7 +30,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_ca</label>
+		<label>Canada</label>
 		<nb_visits>6</nb_visits>
 		<nb_actions>10</nb_actions>
 		<max_actions>3</max_actions>
@@ -59,7 +59,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_fr</label>
+		<label>France</label>
 		<nb_visits>3</nb_visits>
 		<nb_actions>5</nb_actions>
 		<max_actions>3</max_actions>
@@ -88,7 +88,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_id</label>
+		<label>Indonesia</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>3</nb_actions>
 		<max_actions>2</max_actions>
@@ -117,7 +117,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_it</label>
+		<label>Italy</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>4</nb_actions>
 		<max_actions>3</max_actions>
@@ -146,7 +146,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_ti</label>
+		<label>Tibet</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>3</nb_actions>
 		<max_actions>2</max_actions>
@@ -175,7 +175,7 @@
 		<logoHeight>11</logoHeight>
 	</row>
 	<row>
-		<label>UserCountry_country_us</label>
+		<label>United States</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>4</nb_actions>
 		<max_actions>3</max_actions>