--- a/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getRegion_month.xml
+++ b/plugins/UserCountry/tests/System/expected/test_AttributeHistoricalDataWithLocationsTest_testExecute_ShouldReturnLogAfterWorkingWithSomeData__UserCountry.getRegion_month.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <result>
 	<row>
-		<label>General_Unknown</label>
+		<label>Unknown</label>
 		<nb_visits>22</nb_visits>
 		<nb_actions>39</nb_actions>
 		<max_actions>3</max_actions>
@@ -25,12 +25,12 @@
 		<sum_daily_nb_users>0</sum_daily_nb_users>
 		<region>xx</region>
 		<country>xx</country>
-		<country_name>General_Unknown</country_name>
-		<region_name>General_Unknown</region_name>
+		<country_name>Unknown</country_name>
+		<region_name>Unknown</region_name>
 		<logo>plugins/UserCountry/images/flags/xx.png</logo>
 	</row>
 	<row>
-		<label>British Columbia, UserCountry_country_ca</label>
+		<label>British Columbia, Canada</label>
 		<nb_visits>6</nb_visits>
 		<nb_actions>10</nb_actions>
 		<max_actions>3</max_actions>
@@ -55,12 +55,12 @@
 		<segment>regionCode==BC;countryCode==ca</segment>
 		<region>BC</region>
 		<country>ca</country>
-		<country_name>UserCountry_country_ca</country_name>
+		<country_name>Canada</country_name>
 		<region_name>British Columbia</region_name>
 		<logo>plugins/UserCountry/images/flags/ca.png</logo>
 	</row>
 	<row>
-		<label>Franche-Comte, UserCountry_country_fr</label>
+		<label>Franche-Comte, France</label>
 		<nb_visits>3</nb_visits>
 		<nb_actions>5</nb_actions>
 		<max_actions>3</max_actions>
@@ -85,42 +85,12 @@
 		<segment>regionCode==A6;countryCode==fr</segment>
 		<region>A6</region>
 		<country>fr</country>
-		<country_name>UserCountry_country_fr</country_name>
+		<country_name>France</country_name>
 		<region_name>Franche-Comte</region_name>
 		<logo>plugins/UserCountry/images/flags/fr.png</logo>
 	</row>
 	<row>
-		<label>General_Unknown, UserCountry_country_ti</label>
-		<nb_visits>2</nb_visits>
-		<nb_actions>3</nb_actions>
-		<max_actions>2</max_actions>
-		<sum_visit_length>1261</sum_visit_length>
-		<bounce_count>1</bounce_count>
-		<goals>
-			<row idgoal='1'>
-				<nb_conversions>2</nb_conversions>
-				<nb_visits_converted>2</nb_visits_converted>
-				<revenue>10</revenue>
-			</row>
-			<row idgoal='2'>
-				<nb_conversions>1</nb_conversions>
-				<nb_visits_converted>1</nb_visits_converted>
-				<revenue>5</revenue>
-			</row>
-		</goals>
-		<nb_conversions>3</nb_conversions>
-		<revenue>15</revenue>
-		<sum_daily_nb_uniq_visitors>1</sum_daily_nb_uniq_visitors>
-		<sum_daily_nb_users>0</sum_daily_nb_users>
-		<segment>regionCode==1;countryCode==ti</segment>
-		<region>1</region>
-		<country>ti</country>
-		<country_name>UserCountry_country_ti</country_name>
-		<region_name>General_Unknown</region_name>
-		<logo>plugins/UserCountry/images/flags/ti.png</logo>
-	</row>
-	<row>
-		<label>Lazio, UserCountry_country_it</label>
+		<label>Lazio, Italy</label>
 		<nb_visits>2</nb_visits>
 		<nb_actions>4</nb_actions>
 		<max_actions>3</max_actions>
@@ -145,8 +115,38 @@
 		<segment>regionCode==07;countryCode==it</segment>
 		<region>07</region>
 		<country>it</country>
-		<country_name>UserCountry_country_it</country_name>
+		<country_name>Italy</country_name>
 		<region_name>Lazio</region_name>
 		<logo>plugins/UserCountry/images/flags/it.png</logo>
+	</row>
+	<row>
+		<label>Unknown, Tibet</label>
+		<nb_visits>2</nb_visits>
+		<nb_actions>3</nb_actions>
+		<max_actions>2</max_actions>
+		<sum_visit_length>1261</sum_visit_length>
+		<bounce_count>1</bounce_count>
+		<goals>
+			<row idgoal='1'>
+				<nb_conversions>2</nb_conversions>
+				<nb_visits_converted>2</nb_visits_converted>
+				<revenue>10</revenue>
+			</row>
+			<row idgoal='2'>
+				<nb_conversions>1</nb_conversions>
+				<nb_visits_converted>1</nb_visits_converted>
+				<revenue>5</revenue>
+			</row>
+		</goals>
+		<nb_conversions>3</nb_conversions>
+		<revenue>15</revenue>
+		<sum_daily_nb_uniq_visitors>1</sum_daily_nb_uniq_visitors>
+		<sum_daily_nb_users>0</sum_daily_nb_users>
+		<segment>regionCode==1;countryCode==ti</segment>
+		<region>1</region>
+		<country>ti</country>
+		<country_name>Tibet</country_name>
+		<region_name>Unknown</region_name>
+		<logo>plugins/UserCountry/images/flags/ti.png</logo>
 	</row>
 </result>