refs #3611 

When archiving months, use weeks if possible and otherwise days. So instead of fetching and summing 28-31 archives we will fetch and sum only 4-10.

This makes archiving of months much faster. I just tested it on a bit bigger instance. To archive one month for one site the execution time shrank from 450 seconds to 150 seconds. I also did another test with this change as well as the other tweaks that where already merged and the execution time shrank from in total 717s to 104s for 2 months and one site.